Fixed event registration in RCTExNetworkImageManager
Took me a while to firgure out why only loadStart gets invoked, but rest of the events don't work at all.

RCTImageViewManager happens to already registers "loadStart", but since the rest of the event naming are different in this lib (i.e "loadProgress" vs "progress" and so on), they don't work.

Refer to this for reference:

https://github.com/facebook/react-native/blob/61c648d5641da595366801d4a9ef31628cd0eee7/Libraries/Image/RCTImageViewManager.m

This fix makes it work properly on 0.9.0, but won't work on latest >= 0.10.0-rc since they already made a breaking change (from returning NSDictionary to an NSArray), here:

https://github.com/facebook/react-native/blob/master/Libraries/Image/RCTImageViewManager.m

We will need to fix it again if you want to support latest react-native and break compatibility with < 0.10.0.